[BUGFIX] Changed XCLASS recogniation of classes
The XCLASS check did not correctly report when files where used as
XCLASS definition.
Changed this to look into the file which is referenced for the names
of the original and implementation class.
Also the extension key could not be found in these cases.
Added code to check for the extension key in the filepath.